Refactor doc parsing helpers and clean lint warnings
